Introduce injectable audio ingest seam for capture.
Split PortAudio into portaudio_ingest adapter, add TestAudioIngest for
tests, wire Analyzer and Settings through get_audio_ingest(), and keep
AudioBackend() as a backward-compatible alias.

Closes #2
